requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
tomli==2.0.1
